init `product` types file
